backend/requirements.txt
flask
openai
python-dotenv
ffmpeg-python
pydub
gunicorn
requests
